cleanup
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@4721 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ src/player_manager.cpp
@@ -752,7 +752,7 @@ void IPlayerManager::setViewport(const int idx, const sdlx::Rect &rect) {
 
 void IPlayerManager::validateViewports() {
 		if (Map->loaded()) {
-			for(unsigned p = 0; p < _players.size(); ++p) {
+			for(size_t p = 0; p < _players.size(); ++p) {
 				PlayerSlot &slot = _players[p];
 				if (!slot.visible) 
 					continue;
